* Still bugs in propagate ray, reason maybe in meshdata\n * Found bug in surface intersection

+++ include/mesh.h
@@ -33,7 +33,7 @@ typedef Point Vector;
 struct Ray {
   Point p;
   Vector dir;
-  unsigned length;
+  float length;
   
 };
